Migrate barebones Contentful
